refactor: cleanup fork tree utilities.
